FINISHED BABY

+++ app.py
@@ -19,11 +19,12 @@ db.create_all()
 
 USERNAME = "username"
 
+
 @app.get("/")
 def homepage():
     """Redirect to register page"""
 
-    return redirect(url_for('register_user'))
+    return redirect(url_for("register_user"))
 
 
 @app.route("/register", methods=["GET", "POST"])
@@ -31,23 +32,26 @@ def register_user():
     """Register a user to db, re-render form on invalid input
     If valid input, redirect to user profile page
     """
-    if USERNAME in session:
 
-        return redirect(url_for('user_profile', username=session[USERNAME]))
+    if USERNAME in session:
+        return redirect(url_for("user_profile", username=session[USERNAME]))
 
     form = RegisterForm()
 
     if form.validate_on_submit():
         data = {k: v for k, v in form.data.items() if k != "csrf_token"}
 
-        new_user = User.register(**data)
-
-        db.session.add(new_user)
-        db.session.commit()
+        try:
+            new_user = User.register(**data)
+            db.session.add(new_user)
+            db.session.commit()
+        except Exception:
+            flash("Username already taken")
+            return render_template("user_register_form.html", form=form)
 
         session[USERNAME] = new_user.username
 
-        return redirect(url_for('user_profile', username=session[USERNAME]))
+        return redirect(url_for("user_profile", username=session[USERNAME]))
 
     else:
         return render_template("user_register_form.html", form=form)
@@ -58,9 +62,9 @@ def login_user():
     """Show/process login form.
     Ensures user authentication and then directs to user's profile
     """
-    if USERNAME in session:
 
-        return redirect(url_for('user_profile', username=session[USERNAME]))
+    if USERNAME in session:
+        return redirect(url_for("user_profile", username=session[USERNAME]))
 
     form = LoginForm()
 
@@ -71,7 +75,7 @@ def login_user():
 
         if user:
             session[USERNAME] = user.username
-            return redirect(url_for('user_profile', username=session[USERNAME]))
+            return redirect(url_for("user_profile", username=session[USERNAME]))
         else:
             form.username.errors = ["Bad name/password"]
 
@@ -102,22 +106,26 @@ def user_logout():
     else:
         raise Unauthorized()
 
-    return redirect(url_for('homepage'))
+    return redirect(url_for("homepage"))
 
 
-@app.post('/users/<username>/delete')
+@app.post("/users/<username>/delete")
 def delete_user(username):
-    '''Deletes user and all their posts'''
+    """Deletes user and all their posts"""
+
+    if USERNAME not in session or session[USERNAME] != username:
+        raise Unauthorized()
 
     form = CSRFProtectForm()
 
     if form.validate_on_submit():
         user = User.query.get_or_404(username)
-        notes = Note.query.filter(Note.username == username).all()
+        # notes = Note.query.filter(Note.owner_username == username).all()
+        Note.query.filter(Note.owner_username == username).delete()
+        # for note in notes:
+        #     db.session.delete(note)
 
-        for note in notes:
-            db.session.delete(note)
-            session.pop(USERNAME, None)
+        session.pop(USERNAME, None)
 
         db.session.delete(user)
         db.session.commit()
@@ -126,16 +134,15 @@ def delete_user(username):
     else:
         raise Unauthorized()
 
-    return redirect(url_for('homepage'))
+    return redirect(url_for("homepage"))
 
 
 @app.route("/users/<username>/notes/add", methods=["GET", "POST"])
-def add_notes(username):
-    """shows add notes to form and processes input to add to db
-    """
-    if USERNAME not in session or session[USERNAME] != username:
+def add_note(username):
+    """shows add note form and processes input to add to db on submit"""
 
-        return redirect(url_for('homepage'))
+    if USERNAME not in session or session[USERNAME] != username:
+        raise Unauthorized()
 
     form = NoteForm()
 
@@ -147,11 +154,52 @@ def add_notes(username):
         db.session.add(new_note)
         db.session.commit()
 
-        return redirect(url_for('user_profile', username=session[USERNAME]))
+        return redirect(url_for("user_profile", username=session[USERNAME]))
 
     else:
         return render_template("add_note.html", form=form)
 
 
+@app.route("/notes/<note_id>/update", methods=["GET", "POST"])
+def edit_note(note_id):
+    """shows edit note form and processes input to edit existing note on submit"""
+
+    note = Note.query.get_or_404(note_id)
+
+    if USERNAME not in session or session[USERNAME] != note.owner_username:
+        raise Unauthorized()
+
+    form = NoteForm(obj=note)
+
+    if form.validate_on_submit():
+        data = {k: v for k, v in form.data.items() if k != "csrf_token"}
+
+        note.edit_note(**data)
+
+        db.session.commit()
+
+        return redirect(url_for("user_profile", username=session[USERNAME]))
+
+    else:
+        return render_template("edit_note.html", form=form, note=note)
 
 
+@app.post("/notes/<note_id>/delete")
+def delete_note(note_id):
+    """Deletes post, returns to user profile"""
+
+    note = Note.query.get_or_404(note_id)
+
+    if USERNAME not in session or session[USERNAME] != note.owner_username:
+        raise Unauthorized()
+
+    form = CSRFProtectForm()
+
+    if form.validate_on_submit():
+        db.session.delete(note)
+        db.session.commit()
+        flash(f"Note {note.title} deleted.")
+    else:
+        raise Unauthorized()
+
+    return redirect(url_for("user_profile", username=session[USERNAME]))
